split chip components
